+++ src/Endpoint/BulkSignatureEndpoint.php
@@ -8,6 +8,7 @@ use DigitalCz\DigiSign\DigiSign;
 use DigitalCz\DigiSign\Endpoint\Traits\GetEndpointTrait;
 use DigitalCz\DigiSign\Endpoint\Traits\ListEndpointTrait;
 use DigitalCz\DigiSign\Resource\BulkSignature;
+use DigitalCz\DigiSign\Resource\ResourceInterface;
 
 /**
  * @extends ResourceEndpoint<BulkSignature>
@@ -21,5 +22,10 @@ final class BulkSignatureEndpoint extends ResourceEndpoint
     public function __construct(DigiSign $parent)
     {
         parent::__construct($parent, '/api/bulk-signatures', BulkSignature::class);
+    }
+
+    public function resend(string $id): ResourceInterface
+    {
+        return $this->createResource($this->postRequest('/{id}/resend', ['id' => $id]));
     }
 }